useGraph gets createSceneMeshes
